Adding a git shell include file.
I have been thinking the code is becoming too spaghetti like recently
as I add on new functionality.  I knew this day was coming, so we I
am building an include structure that houses various aspects of the
Git / Drush interaction.  This makes it more modular, which is always
good.

The first of these installments is a git.shell.inc file that houses
all of the shell wrappers used in the git commands.  These make it
easier because we do not have to worry about keeping Git context
when working with multiple repositories, such as with Git submodules.
And they just generally make it easier to work with Git through the
PHP and Drush shell interface.